% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutline.r
\name{detect.outline}
\alias{detect.outline}
\alias{outline.detect}
\title{detect.outline}
\usage{
detect.outline(input.file = NULL, image.grey = NULL,
  grey.mode = "normal.grey", parameter.for.end = NULL, distance = 20,
  parameter.distance.deviation = 2, image.information = NULL,
  image = NULL)
}
\arguments{
\item{input.file}{A character:Full path of an image that shall be used for
outline detection}

\item{image.grey}{A matrix of numbers: Grey values of the otolith image
(values between 0 and 1)}

\item{grey.mode}{A character: Determines how the grey image shall be
calculated}

\item{parameter.for.end}{A number: The edge is being found by calculating
a floating mean. The function marks an edge if the next point differs
more than floating mean divided by parameter.for.end}

\item{distance}{A number: Lines for finding the outline of the otolith
are this many pixels apart on the x- or y-axes}

\item{parameter.distance.deviation}{A number: Outliers get detected by
multiplying the parameter with the mean distance between points}

\item{image.information}{A matrix of numbers: Contains additional pixel
information of the image (values between 0 and 1)}

\item{image}{A matrix of numbers: Contains the information of the otolith
image}
}
\description{
Detect the outline of the otolith
}
\details{
We will add more later...To be continued
}
\author{
Kai Budde
}
