% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editImage.r
\name{edit.image}
\alias{edit.image}
\alias{image.edit}
\title{edit.image}
\usage{
\method{edit}{image}(image = NULL, file.name = NULL, grey.mode = NULL,
  brightness = NULL, contrast = NULL, red = NULL, green = NULL,
  blue = NULL)
}
\arguments{
\item{image}{An array of numbers between 0 and 1}

\item{file.name}{A character (location of the image)}

\item{grey.mode}{A character}

\item{brightness}{A number}

\item{contrast}{A character}

\item{red}{A character}

\item{green}{A character}

\item{blue}{A character}
}
\description{
Adjust color, brightness or contrast of the image
}
\details{
By using an x-y-3(rgb)-representation of an image, you can
change the color, brightness or contrast of that image. For adjusting
brightness or contrast use a number between -1 and 1 as parameter.
To get more intense colors use a number between 0 and 1 for each color
parameter. By entering grey.mode = TRUE one will get a normal grey image.
Other options for grey.mode are "luminosity.grey", red.grey",
"green.grey" and "blue.grey".
}
\author{
Kai Budde
}
