% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMidpoint.r
\name{find.midpoint}
\alias{find.midpoint}
\alias{findmidpoint}
\alias{find.Midpoint}
\title{find.midpoint}
\usage{
find.midpoint(image = NULL, image.grey = NULL, center.point = NULL,
  number.of.blocks.in.row = NULL, search.length = 180, image.path)
}
\arguments{
\item{image}{A matrix of numbers: Contains the information of the otolith
image}

\item{image.grey}{A matrix of numbers: Grey values of the otolith image
(values between 0 and 1)}

\item{center.point}{A number: Coordinates of the starting point for the
box to look for the darkest point}

\item{number.of.blocks.in.row}{A number: Number of blocks in a row or
column (the search field will then be number x number big)}

\item{search.length}{A number: Pixel length of the entire search field.}
}
\description{
Find midpoint of otolith close to intersection of two lines
}
\details{
We will add more later...To be continued
}
\author{
Kai Budde
}
