% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAge.r
\name{read.age}
\alias{read.age}
\alias{readage}
\alias{age.read}
\alias{ageread}
\title{read.age}
\usage{
read.age(input.dir = NULL, input.file = NULL, user.file.en = NULL,
  user.file.de = NULL, distance = 20, parameter.for.end = NULL,
  parameter.distance.deviation = 2, par.hyaline = 1.1,
  points.to.jump = 10, points.for.reconnecting = 20,
  remove.points.at.end = 25, remove.points.at.beginning = 0, t = 2,
  difference.in.percent = 0.25, outer.zone.pixel = 50,
  show.all.possible.rings = FALSE, distance.second.lines = 50,
  points.to.look.for.midpoint = 50,
  find.midpoint.method = "intersectionMidpointThree")
}
\arguments{
\item{input.dir}{A character}

\item{input.file}{A character}

\item{user.file.en}{A character}

\item{user.file.de}{A character}

\item{distance}{A number (Lines for finding the outline of the otolith
are this many pixels apart on the x- and y-axes.)}

\item{parameter.for.end}{A number (The edge is being found by calculating
a floating mean. The function marks an edge, if the next point differs
more than floating mean divided by parameter.for.end.)}

\item{parameter.distance.deviation}{A number (Outliers get detected by
multiplying the parameter with the mean distance between points.)}

\item{par.hyaline}{A number (This parameter is multiplied by the mean
of the brightness of the image to give a lower limit for ring detection.)}

\item{points.to.jump}{A number (It indicates how many points can be
missed while finding a large ring.)}

\item{points.for.reconnecting}{A number (It indicates how far rings
should be away from each other. Otherwise they are being merged.)}

\item{remove.points.at.end}{A number (It indicates how many points should
be disregarded at the end of the line.)}

\item{remove.points.at.beginning}{A number (It indicates how far the
first ring should be away from the starting (midpoint).)}

\item{t}{A number (This parameter is being multiplied by the standard
deviation in order to delete outliers in the line values.)}

\item{difference.in.percent}{A number (Rings that are smaller than this
parameter in comparison to the previous and next ring are being deleted.)}

\item{outer.zone.pixel}{A number (It defines how many pixels at the
end of the line belong to the outer zone.)}

\item{show.all.possible.rings}{A boolean (If it is true, than all
possible rings are being maked blue.)}

\item{find.midpoint.method}{A character ()}
}
\description{
Main function for determining the age of a fish
}
\details{
Input should be tif-format.
The output will be written in the current directory.
Please be aware that the coordinate system is turned by 90° to the right.
The origin of the x- and y-axes is in the upper-left corner of the image.
(The x-axis points downwards and the y-axis to the right.)
}
\author{
Kai Budde
}
