% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutline.r
\name{detect.outline}
\alias{detect.outline}
\alias{outline.detect}
\title{detect.outline}
\usage{
detect.outline(image.grey = NULL, image.information = NULL, distance = 20,
  parameter.for.end = NULL, parameter.distance.deviation = 2,
  file.name = NULL, grey.mode = "normal.grey")
}
\arguments{
\item{image.grey}{An array (x,y) of values between 0 and 1}

\item{image.information}{An array (x,y) of values between 0 and 1}

\item{distance}{A number (Lines for finding the outline of the otolith
are this many pixels apart on the x- and y-axes.)}

\item{parameter.for.end}{A number (The edge is being found by calculating
a floating mean. The function marks an edge, if the next point differs
more than floating mean divided by parameter.for.end.)}

\item{parameter.distance.deviation}{A number (Outliers get detected by
multiplying the parameter with the mean distance between points.)}

\item{file.name}{A character}

\item{grey.mode}{A character}
}
\description{
Detect the outline of the otolith
}
\details{
We will add more later...To be continued
}
\author{
Kai Budde
}
